Dictionaries: changed Forcellini to non-digitised version, removed Notre-Dame, Numen URL uses text fragment, updated TLL URL
Explanations:
Digitised version of Forcellini is not yet complete, so the non-digitised version is more comprehensive for the time being.
Notre-Dame no longer works as an interface to Whitaker’s Words.
I don’t actually have access to TLL since I’m not a student, but I believe the URL is correct!